Cases

Processor

process :: CaseTransformer -> Delimiter -> Text -> Text

Extract separate words from an arbitrary text using a smart parser and produce a new text using case transformation and delimiter functions.

Note: to skip case transformation use the id function.

Case Transformers

type CaseTransformer = Part -> Part

lower :: CaseTransformer

upper :: CaseTransformer

title :: CaseTransformer

Delimiters

type Delimiter = Folder (Maybe Text)

spinal :: Delimiter

snake :: Delimiter

camel :: Delimiter

Default Processors

spinalize :: Text -> Text

Transform an arbitrary text into a lower spinal case.

Same as (process lower spinal).

snakify :: Text -> Text

Transform an arbitrary text into a lower snake case.

Same as (process lower snake).

camelize :: Text -> Text

Transform an arbitrary text into a camel case, while preserving the case of the first character.

Same as (process id camel).